+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [7.0.0-beta.4] - 2023-05-24
+
 ### Fixed
 
 - Added `a-input__wrapper--inline` class for inline `TextField` and `TextArea` component
@@ -17,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `locale` props for `Calendar` (and Datepicker via `calendarProps`)
+
 
 ## [7.0.0-beta.3] - 2023-05-08
 
@@ -36,9 +39,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `disabled` prop for `Button`
 
+
 ## [7.0.0-beta.2] - 2023-05-08
 
 Unpublished broken version
+
 
 ## [7.0.0-beta.1] - 2023-04-20
 
@@ -64,11 +69,12 @@ Unpublished broken version
 - `multiple` option for `Autocomplete`
 - A warning is logged for a wrong or unknown `Icon`
 
+
 ## [7.0.0-beta.0] - 2023-03-30
 
 ### Changed
 
-- New React components for Core Branding 6.2.2
+- New React components for Core Branding 6.3.0
 
 ### Removed